+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added @Required decorator for entity properties
 * Reworked how the @Validate object schema works, [see the docs](https://github.com/tudddorrr/koa-rest-services/tree/main/docs/validation.md)
 * Removed the @Before and @After decorators
+* Move registered services into `ctx.state`
 
 ## 4.0.3 (15th November 2021)
 * Only reject null or undefined values in `@Validate` schemas